V2.2.4.0.0
This is an F# functional first units library. It provides an API based on core units concepts and is built around the concepts of functional first design and type safety. From this philosophy, this library provides a way of maintaining unit conversions and manipulations which allows you to focus on the hard work without being burdened with the though of having a unit discrepancy.